refactor(Perceptions): move to separate file
