+++ src/main/scala/ulisse/entities/simulation/SimulationAgent.scala
@@ -1,6 +1,6 @@
 package ulisse.entities.simulation
 
-import ulisse.entities.simulation.Environments.{Perception, RailwayEnvironment}
+import ulisse.entities.simulation.Environments.RailwayEnvironment
 import ulisse.entities.simulation.Simulations.Actions.SimulationAction
 import ulisse.entities.train.Trains.Train
